# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Network Asset Discovery & NAC Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all devices
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/DeviceStatus"
        - in: query
          name: vlanId
          schema:
            type: integer
        - in: query
          name: siteId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated device list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDevices"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Manually register a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Update device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /devices/{id}/approve:
    post:
      operationId: approveDevice
      tags: [devices]
      summary: Approve a device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceApproval"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/reject:
    post:
      operationId: rejectDevice
      tags: [devices]
      summary: Reject a device
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRejection"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/quarantine:
    post:
      operationId: quarantineDevice
      tags: [devices]
      summary: Move device to quarantine
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quarantined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/history:
    get:
      operationId: getDeviceHistory
      tags: [devices]
      summary: Get device audit history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceHistory"
  /devices/bulk-approve:
    post:
      operationId: bulkApproveDevices
      tags: [devices]
      summary: Bulk approve devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkActionRequest"
      responses:
        "200":
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
  /devices/bulk-reject:
    post:
      operationId: bulkRejectDevices
      tags: [devices]
      summary: Bulk reject devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkActionRequest"
      responses:
        "200":
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
  /devices/bulk-vlan:
    post:
      operationId: bulkAssignVlan
      tags: [devices]
      summary: Bulk assign VLAN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkVlanAssignment"
      responses:
        "200":
          description: Bulk result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"

  # ───────────────────────────── DISCOVERY ─────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - in: query
          name: severity
          schema:
            type: string
            enum: [critical, warning, info]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlerts"
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    DeviceApproval:
      type: object
      properties:
        vlanId:
          type: integer
    DeviceRejection:
      type: object
      properties:
        reason:
          type: string
    BulkVlanAssignment:
      type: object
      required: [ids, vlanId]
      properties:
        ids:
          type: array
          items:
            type: integer
        vlanId:
          type: integer
...
    DeviceStatus:
      type: string
      enum: [DISCOVERED, PENDING, APPROVED, REJECTED, QUARANTINED, SYNCED]
    Device:
      type: object
      properties:
        id:
          type: integer
        macAddress:
          type: string
        ipAddress:
          type: string
          nullable: true
        hostname:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        vendor:
          type: string
          nullable: true
        operatingSystem:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        switchName:
          type: string
          nullable: true
        switchPort:
          type: string
          nullable: true
        vlanId:
          type: integer
          nullable: true
        vlan:
          $ref: "#/components/schemas/Vlan"
          nullable: true
        siteId:
          type: integer
          nullable: true
        site:
          $ref: "#/components/schemas/Site"
          nullable: true
        status:
          $ref: "#/components/schemas/DeviceStatus"
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
          nullable: true
        approvedBy:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        radiusSynced:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - macAddress
        - status
        - firstSeen
        - lastSeen
        - radiusSynced
        - createdAt
        - updatedAt
    CreateDeviceRequest:
      type: object
      required: [macAddress]
      properties:
        macAddress:
          type: string
        ipAddress:
          type: string
        hostname:
          type: string
        username:
          type: string
        vendor:
          type: string
        operatingSystem:
          type: string
        department:
          type: string
        switchName:
          type: string
        switchPort:
          type: string
        vlanId:
          type: integer
        siteId:
          type: integer
        notes:
          type: string
    UpdateDeviceRequest:
      type: object
      properties:
        hostname:
          type: string
        username:
          type: string
        vendor:
          type: string
        operatingSystem:
          type: string
        department:
          type: string
        vlanId:
          type: integer
        siteId:
          type: integer
        notes:
          type: string
    PaginatedDevices:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DeviceHistory:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        action:
          type: string
        oldStatus:
          type: string
          nullable: true
        newStatus:
          type: string
          nullable: true
        oldVlan:
          type: string
          nullable: true
        newVlan:
          type: string
          nullable: true
        performedBy:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, deviceId, action, createdAt]
    BulkActionRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
        vlanId:
          type: integer
        reason:
          type: string
    BulkActionResult:
      type: object
      required: [succeeded, failed]
      properties:
        succeeded:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Vlan:
      type: object
      required: [id, vlanId, name]
      properties:
        id:
          type: integer
        vlanId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [production, quarantine, guest, management]
          nullable: true
        isQuarantine:
          type: boolean
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Site:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [site, branch, building, floor, rack]
        deviceCount:
          type: integer
        switchCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Alert:
      type: object
      required: [id, type, severity, status, message, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_device, device_offline, duplicate_mac, duplicate_ip, port_change, unauthorized_device, policy_violation]
        severity:
          type: string
          enum: [critical, warning, info]
        status:
          type: string
          enum: [open, acknowledged, resolved]
        message:
          type: string
        deviceId:
          type: integer
          nullable: true
        device:
          $ref: "#/components/schemas/Device"
          nullable: true
        acknowledgedBy:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PaginatedAlerts:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...